Created logic for posting reports for today - DailyReportsService. Created calculating logic for 'prosuto' - ProsutoService. Added custom convert from float to numeric type in db.

+++ backend/OLDBRICK-STANJE-ARTIKALA-APP/OLDBRICK-STANJE-ARTIKALA-APP/DTOs/DailyReports/DailyReportResponseDto.cs
@@ -4,6 +4,6 @@
     {
         public int IdNaloga { get; set; }
         public DateOnly Datum { get; set; }
-        public decimal Prosuto { get; set; }
+        public float Prosuto { get; set; }
     }
 }
